@@ -15,11 +15,10 @@
     "@ant-design/icons": "^4.7.0",
     "@types/node": "^16.11.6",
     "@types/react": "^18.2.14",
-    "@types/react-dom": "^17.0.0",
+    "@types/react-dom": "^18.2.18",
     "antd": "^5.6.4",
     "axios": "^0.26.1",
     "classnames": "^2.3.1",
-    "craco-sass-resources-loader": "^1.1.0",
     "crypto-js": "^4.2.0",
     "dayjs": "^1.11.10",
     "history": "^5.3.0",
@@ -28,22 +27,22 @@
     "lint-staged": "^15.2.0",
     "lodash": "^4.17.21",
     "normalize.css": "^8.0.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "^5.0.1",
     "sass": "^1.41.0",
-    "sass-resources-loader": "^2.2.5",
     "typescript": "^5.3.3",
     "zustand": "^4.4.7"
   },
   "scripts": {
     "prepare": "husky install",
-    "dev": "cross-env  NODE_ENV=develop craco start ",
-    "build": "cross-env  NODE_ENV=production craco build",
     "lint": "eslint --fix 'src/**/*.{js,jsx,ts,tsx,json,md}'",
     "format": "prettier --write 'src/**/*.{js,jsx,ts,tsx,json,md}'",
-    "lint-staged": "tsc && lint-staged"
+    "lint-staged": "tsc && lint-staged",
+    "dev": "vite --mode develop",
+    "build": "tsc && vite build --mode production",
+    "preview": "vite preview"
   },
   "browserslist": {
     "production": [
@@ -82,8 +81,6 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "husky": "^8.0.3",
-    "postcss": "8",
-    "postcss-scss": "^4.0.3",
     "prettier": "^3.1.1",
     "react-app-rewire-hot-loader": "^2.0.1",
     "react-hot-loader": "^4.13.0",
